requirements.txt
biopython==1.72
bleach==1.5.0
brewer2mpl==1.4.1
click==6.7
colorama==0.3.9
cycler==0.10.0
decorator==4.3.0
Django==2.0
docopt==0.6.2
future==0.17.0
ggplot==0.11.5
graphviz==0.9
hpsklearn==0.0.3
html5lib==0.9999999
hyperopt==0.1.1
Markdown==2.6.9
matplotlib==2.1.1
mysql-connector==2.1.7
mysqlclient==1.3.10
networkx==2.2
nose==1.3.7
numpy==1.15.4
pandas==0.23.4
patsy==0.4.1
Pillow==5.2.0
progressbar2==3.34.2
protobuf==3.4.0
PubChemPy==1.0.4
pullenrichment==0.0.0
pymongo==3.7.2
PyMySQL==0.7.11
pyparsing==2.2.0
python-dateutil==2.7.5
python-utils==2.2.0
pytz==2018.7
PyYAML==3.12
scikit-learn==0.19.1
scipy==1.3.0
seaborn==0.9.0
six==1.12.0
sklearn==0.0
SQLAlchemy==1.1.13
statsmodels==0.9.0
tensorflow==1.3.0
tensorflow-tensorboard==0.1.8
unicodecsv==0.14.1
Werkzeug==0.12.2
wordcloud==1.5.0
xgbmagic==0.0.3.2
xgboost==0.80
XGBoost-Ranking==0.7.1
xlrd==1.1.0
